+++ src/TriggerPrimitive.hh
@@ -1,9 +1,10 @@
 #pragma once
+#include <cstdint>
 
 namespace DuneTriggers{
   struct TriggerPrimitive {
-    int time_start = {0};
-    int time_peak = {0};
+    int64_t time_start = {0};
+    int64_t time_peak = {0};
     int time_over_threshold = {0};
     int channel = {0};
     int adc_integral = {0};
